WFLY-15169 Refactor EJB client marshallers into a separate module
