+++ clustering/ejb/client/src/main/java/org/wildfly/clustering/ejb/client/SessionIDSerializer.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.wildfly.clustering.ejb.infinispan;
+package org.wildfly.clustering.ejb.client;
 
 import java.io.DataInput;
 import java.io.DataOutput;
