+++ clustering/ejb/client/src/main/resources/org.jboss.ejb.client.proto
