+++ clustering/ejb/infinispan/src/main/java/org/wildfly/clustering/ejb/infinispan/group/BeanGroupSerializationContextInitializer.java
@@ -25,7 +25,7 @@ package org.wildfly.clustering.ejb.infinispan.group;
 import java.util.Map;
 
 import org.infinispan.protostream.SerializationContext;
-import org.wildfly.clustering.ejb.infinispan.EJBClientMarshallingProvider;
+import org.jboss.ejb.client.SessionID;
 import org.wildfly.clustering.marshalling.protostream.AbstractSerializationContextInitializer;
 import org.wildfly.clustering.marshalling.protostream.FunctionalMarshaller;
 import org.wildfly.clustering.marshalling.spi.ByteBufferMarshalledValue;
@@ -40,7 +40,7 @@ public class BeanGroupSerializationContextInitializer extends AbstractSerializat
     @SuppressWarnings("unchecked")
     @Override
     public void registerMarshallers(SerializationContext context) {
-        context.registerMarshaller(new FunctionalMarshaller<>(InfinispanBeanGroupKey.class, EJBClientMarshallingProvider.SESSION_ID, InfinispanBeanGroupKey::getId, InfinispanBeanGroupKey::new));
+        context.registerMarshaller(new FunctionalMarshaller<>((Class<InfinispanBeanGroupKey<SessionID>>) (Class<?>) InfinispanBeanGroupKey.class, SessionID.class, InfinispanBeanGroupKey::getId, InfinispanBeanGroupKey::new));
         context.registerMarshaller(new FunctionalMarshaller<>(InfinispanBeanGroupEntry.class, (Class<MarshalledValue<Map<Object, Object>, Object>>) (Class<?>) ByteBufferMarshalledValue.class, InfinispanBeanGroupEntry<Object, Object, Object>::getBeans, InfinispanBeanGroupEntry<Object, Object, Object>::new));
     }
 }
